Move `contexts.js` to tests.
